pipeline/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0
mediapipe>=0.8.0
ultralytics>=8.0.0
openai>=1.0.0
pyyaml>=6.0
aiohttp>=3.8.0
aiofiles>=0.8.0
tqdm>=4.62.0
tenacity>=8.0.0
pillow>=8.0.0
ffmpeg-python>=0.2.0 